landforge_verify_document/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "landforge_verify_document"
version = "0.1.0"
dependencies = [
  "nvidia-nat[langchain]~=1.5",
]
requires-python = ">=3.11,<3.14"
description = "Custom NeMo Agent Toolkit Workflow"
classifiers = ["Programming Language :: Python"]



[project.entry-points.'nat.components']
landforge_verify_document = "landforge_verify_document.register"
